requirements.txt
absl-py==0.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
google-auth==1.7.0
google-auth-oauthlib==0.4.1
grpcio==1.25.0
idna==2.8
joblib==0.14.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.3
oauthlib==3.1.0
opencv-contrib-python==4.1.1.26
opencv-torchvision-transforms-yuzhiyang==0.0.1
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
tensorboard==2.0.1
torch==1.3.1
torchvision==0.4.2
urllib3==1.25.7
Werkzeug==0.16.0
